Combine RequirementsHelper and UpdateHelper
